requirements.txt
attrs==21.4.0
black==21.12b0
cfgv==3.3.1
click==8.0.3
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
identify==2.4.4
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-mock==3.6.1
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
virtualenv==20.13.0
zipp==3.7.0
